separate operations tests into packages (#4911)
operations integration tets are outgrowing the time limits on golang
package runtime. This split is to help individual test suites stay
within the 20 minute limit.

No code changes.  Only movement/renaming.
