+++ src/internal/operations/test/m365/groups/groups_test.go
@@ -1,4 +1,4 @@
-package test_test
+package groups_test
 
 import (
 	"context"
@@ -9,6 +9,7 @@ import (
 
 	evmock "github.com/alcionai/corso/src/internal/events/mock"
 	"github.com/alcionai/corso/src/internal/m365/collection/drive"
+	. "github.com/alcionai/corso/src/internal/operations/test/m365"
 	"github.com/alcionai/corso/src/internal/tester"
 	"github.com/alcionai/corso/src/internal/tester/tconfig"
 	"github.com/alcionai/corso/src/internal/version"
@@ -24,7 +25,7 @@ import (
 
 type GroupsBackupIntgSuite struct {
 	tester.Suite
-	its intgTesterSetup
+	its IntgTesterSetup
 }
 
 func TestGroupsBackupIntgSuite(t *testing.T) {
@@ -36,18 +37,18 @@ func TestGroupsBackupIntgSuite(t *testing.T) {
 }
 
 func (suite *GroupsBackupIntgSuite) SetupSuite() {
-	suite.its = newIntegrationTesterSetup(suite.T())
+	suite.its = NewIntegrationTesterSetup(suite.T())
 }
 
 func (suite *GroupsBackupIntgSuite) TestBackup_Run_groups() {
 	var (
-		resourceID = suite.its.group.ID
+		resourceID = suite.its.Group.ID
 		sel        = selectors.NewGroupsBackup([]string{resourceID})
 	)
 
 	sel.Include(selTD.GroupsBackupLibraryFolderScope(sel))
 
-	runBasicDriveishBackupTests(
+	RunBasicDriveishBackupTests(
 		suite,
 		path.GroupsService,
 		control.DefaultOptions(),
@@ -60,13 +61,13 @@ func (suite *GroupsBackupIntgSuite) TestBackup_Run_incrementalGroups() {
 
 func (suite *GroupsBackupIntgSuite) TestBackup_Run_extensionsGroups() {
 	var (
-		resourceID = suite.its.group.ID
+		resourceID = suite.its.Group.ID
 		sel        = selectors.NewGroupsBackup([]string{resourceID})
 	)
 
 	sel.Include(selTD.GroupsBackupLibraryFolderScope(sel))
 
-	runDriveishBackupWithExtensionsTests(
+	RunDriveishBackupWithExtensionsTests(
 		suite,
 		path.GroupsService,
 		control.DefaultOptions(),
@@ -79,7 +80,7 @@ func (suite *GroupsBackupIntgSuite) TestBackup_Run_extensionsGroups() {
 
 type GroupsBackupTreeIntgSuite struct {
 	tester.Suite
-	its intgTesterSetup
+	its IntgTesterSetup
 }
 
 func TestGroupsBackupTreeIntgSuite(t *testing.T) {
@@ -91,12 +92,12 @@ func TestGroupsBackupTreeIntgSuite(t *testing.T) {
 }
 
 func (suite *GroupsBackupTreeIntgSuite) SetupSuite() {
-	suite.its = newIntegrationTesterSetup(suite.T())
+	suite.its = NewIntegrationTesterSetup(suite.T())
 }
 
 func (suite *GroupsBackupTreeIntgSuite) TestBackup_Run_treeGroups() {
 	var (
-		resourceID = suite.its.group.ID
+		resourceID = suite.its.Group.ID
 		sel        = selectors.NewGroupsBackup([]string{resourceID})
 		opts       = control.DefaultOptions()
 	)
@@ -105,7 +106,7 @@ func (suite *GroupsBackupTreeIntgSuite) TestBackup_Run_treeGroups() {
 
 	opts.ToggleFeatures.UseDeltaTree = true
 
-	runBasicDriveishBackupTests(
+	RunBasicDriveishBackupTests(
 		suite,
 		path.GroupsService,
 		opts,
@@ -121,7 +122,7 @@ func (suite *GroupsBackupTreeIntgSuite) TestBackup_Run_treeIncrementalGroups() {
 
 func (suite *GroupsBackupTreeIntgSuite) TestBackup_Run_treeExtensionsGroups() {
 	var (
-		resourceID = suite.its.group.ID
+		resourceID = suite.its.Group.ID
 		sel        = selectors.NewGroupsBackup([]string{resourceID})
 		opts       = control.DefaultOptions()
 	)
@@ -130,7 +131,7 @@ func (suite *GroupsBackupTreeIntgSuite) TestBackup_Run_treeExtensionsGroups() {
 
 	opts.ToggleFeatures.UseDeltaTree = true
 
-	runDriveishBackupWithExtensionsTests(
+	RunDriveishBackupWithExtensionsTests(
 		suite,
 		path.GroupsService,
 		opts,
@@ -143,10 +144,10 @@ func (suite *GroupsBackupTreeIntgSuite) TestBackup_Run_treeExtensionsGroups() {
 
 func runGroupsIncrementalBackupTests(
 	suite tester.Suite,
-	its intgTesterSetup,
+	its IntgTesterSetup,
 	opts control.Options,
 ) {
-	sel := selectors.NewGroupsRestore([]string{its.group.ID})
+	sel := selectors.NewGroupsRestore([]string{its.Group.ID})
 
 	ic := func(cs []string) selectors.Selector {
 		sel.Include(sel.LibraryFolders(cs, selectors.PrefixMatch()))
@@ -157,25 +158,25 @@ func runGroupsIncrementalBackupTests(
 		t *testing.T,
 		ctx context.Context,
 	) string {
-		return its.group.RootSite.DriveID
+		return its.Group.RootSite.DriveID
 	}
 
 	gtsi := func(
 		t *testing.T,
 		ctx context.Context,
 	) string {
-		return its.group.RootSite.ID
+		return its.Group.RootSite.ID
 	}
 
 	grh := func(ac api.Client) drive.RestoreHandler {
 		return drive.NewSiteRestoreHandler(ac, path.GroupsService)
 	}
 
-	runIncrementalDriveishBackupTest(
+	RunIncrementalDriveishBackupTest(
 		suite,
 		opts,
-		its.group.ID,
-		its.user.ID,
+		its.Group.ID,
+		its.User.ID,
 		path.GroupsService,
 		path.LibrariesCategory,
 		ic,
@@ -193,7 +194,7 @@ func (suite *GroupsBackupIntgSuite) TestBackup_Run_groups9VersionBumpBackup() {
 
 	var (
 		mb      = evmock.NewBus()
-		sel     = selectors.NewGroupsBackup([]string{suite.its.group.ID})
+		sel     = selectors.NewGroupsBackup([]string{suite.its.Group.ID})
 		opts    = control.DefaultOptions()
 		whatSet = deeTD.CategoryFromRepoRef
 	)
@@ -203,7 +204,7 @@ func (suite *GroupsBackupIntgSuite) TestBackup_Run_groups9VersionBumpBackup() {
 		selTD.GroupsBackupChannelScope(sel),
 		sel.Conversation(selectors.Any()))
 
-	bo, bod := prepNewTestBackupOp(
+	bo, bod := PrepNewTestBackupOp(
 		t,
 		ctx,
 		mb,
@@ -211,41 +212,41 @@ func (suite *GroupsBackupIntgSuite) TestBackup_Run_groups9VersionBumpBackup() {
 		opts,
 		version.All8MigrateUserPNToID,
 		count.New())
-	defer bod.close(t, ctx)
+	defer bod.Close(t, ctx)
 
-	runAndCheckBackup(t, ctx, &bo, mb, false)
-	checkBackupIsInManifests(
+	RunAndCheckBackup(t, ctx, &bo, mb, false)
+	CheckBackupIsInManifests(
 		t,
 		ctx,
-		bod.kw,
-		bod.sw,
+		bod.KW,
+		bod.SW,
 		&bo,
-		bod.sel,
-		bod.sel.ID(),
+		bod.Sel,
+		bod.Sel.ID(),
 		path.ChannelMessagesCategory)
 
 	_, expectDeets := deeTD.GetDeetsInBackup(
 		t,
 		ctx,
 		bo.Results.BackupID,
-		bod.acct.ID(),
-		bod.sel.ID(),
+		bod.Acct.ID(),
+		bod.Sel.ID(),
 		path.GroupsService,
 		whatSet,
-		bod.kms,
-		bod.sss)
+		bod.KMS,
+		bod.SSS)
 	deeTD.CheckBackupDetails(
 		t,
 		ctx,
 		bo.Results.BackupID,
 		whatSet,
-		bod.kms,
-		bod.sss,
+		bod.KMS,
+		bod.SSS,
 		expectDeets,
 		false)
 
 	mb = evmock.NewBus()
-	forcedFull := newTestBackupOp(
+	forcedFull := NewTestBackupOp(
 		t,
 		ctx,
 		bod,
@@ -254,34 +255,34 @@ func (suite *GroupsBackupIntgSuite) TestBackup_Run_groups9VersionBumpBackup() {
 		count.New())
 	forcedFull.BackupVersion = version.Groups9Update
 
-	runAndCheckBackup(t, ctx, &forcedFull, mb, false)
-	checkBackupIsInManifests(
+	RunAndCheckBackup(t, ctx, &forcedFull, mb, false)
+	CheckBackupIsInManifests(
 		t,
 		ctx,
-		bod.kw,
-		bod.sw,
+		bod.KW,
+		bod.SW,
 		&forcedFull,
-		bod.sel,
-		bod.sel.ID(),
+		bod.Sel,
+		bod.Sel.ID(),
 		path.ChannelMessagesCategory)
 
 	_, expectDeets = deeTD.GetDeetsInBackup(
 		t,
 		ctx,
 		forcedFull.Results.BackupID,
-		bod.acct.ID(),
-		bod.sel.ID(),
+		bod.Acct.ID(),
+		bod.Sel.ID(),
 		path.GroupsService,
 		whatSet,
-		bod.kms,
-		bod.sss)
+		bod.KMS,
+		bod.SSS)
 	deeTD.CheckBackupDetails(
 		t,
 		ctx,
 		forcedFull.Results.BackupID,
 		whatSet,
-		bod.kms,
-		bod.sss,
+		bod.KMS,
+		bod.SSS,
 		expectDeets,
 		false)
 
@@ -303,7 +304,7 @@ func (suite *GroupsBackupIntgSuite) TestBackup_Run_groupsBasic() {
 	var (
 		mb      = evmock.NewBus()
 		counter = count.New()
-		sel     = selectors.NewGroupsBackup([]string{suite.its.group.ID})
+		sel     = selectors.NewGroupsBackup([]string{suite.its.Group.ID})
 		opts    = control.DefaultOptions()
 		whatSet = deeTD.CategoryFromRepoRef
 	)
@@ -313,37 +314,37 @@ func (suite *GroupsBackupIntgSuite) TestBackup_Run_groupsBasic() {
 		selTD.GroupsBackupChannelScope(sel),
 		sel.Conversation(selectors.Any()))
 
-	bo, bod := prepNewTestBackupOp(t, ctx, mb, sel.Selector, opts, version.Backup, counter)
-	defer bod.close(t, ctx)
+	bo, bod := PrepNewTestBackupOp(t, ctx, mb, sel.Selector, opts, version.Backup, counter)
+	defer bod.Close(t, ctx)
 
-	runAndCheckBackup(t, ctx, &bo, mb, false)
-	checkBackupIsInManifests(
+	RunAndCheckBackup(t, ctx, &bo, mb, false)
+	CheckBackupIsInManifests(
 		t,
 		ctx,
-		bod.kw,
-		bod.sw,
+		bod.KW,
+		bod.SW,
 		&bo,
-		bod.sel,
-		bod.sel.ID(),
+		bod.Sel,
+		bod.Sel.ID(),
 		path.ChannelMessagesCategory)
 
 	_, expectDeets := deeTD.GetDeetsInBackup(
 		t,
 		ctx,
 		bo.Results.BackupID,
-		bod.acct.ID(),
-		bod.sel.ID(),
+		bod.Acct.ID(),
+		bod.Sel.ID(),
 		path.GroupsService,
 		whatSet,
-		bod.kms,
-		bod.sss)
+		bod.KMS,
+		bod.SSS)
 	deeTD.CheckBackupDetails(
 		t,
 		ctx,
 		bo.Results.BackupID,
 		whatSet,
-		bod.kms,
-		bod.sss,
+		bod.KMS,
+		bod.SSS,
 		expectDeets,
 		false)
 }
@@ -357,30 +358,30 @@ func (suite *GroupsBackupIntgSuite) TestBackup_Run_groupsExtensions() {
 	var (
 		mb      = evmock.NewBus()
 		counter = count.New()
-		sel     = selectors.NewGroupsBackup([]string{suite.its.group.ID})
+		sel     = selectors.NewGroupsBackup([]string{suite.its.Group.ID})
 		opts    = control.DefaultOptions()
 		tenID   = tconfig.M365TenantID(t)
 		svc     = path.GroupsService
 		ws      = deeTD.DriveIDFromRepoRef
 	)
 
-	opts.ItemExtensionFactory = getTestExtensionFactories()
+	opts.ItemExtensionFactory = GetTestExtensionFactories()
 
 	// does not apply to channel messages
 	sel.Include(selTD.GroupsBackupLibraryFolderScope(sel))
 
-	bo, bod := prepNewTestBackupOp(t, ctx, mb, sel.Selector, opts, version.Backup, counter)
-	defer bod.close(t, ctx)
+	bo, bod := PrepNewTestBackupOp(t, ctx, mb, sel.Selector, opts, version.Backup, counter)
+	defer bod.Close(t, ctx)
 
-	runAndCheckBackup(t, ctx, &bo, mb, false)
-	checkBackupIsInManifests(
+	RunAndCheckBackup(t, ctx, &bo, mb, false)
+	CheckBackupIsInManifests(
 		t,
 		ctx,
-		bod.kw,
-		bod.sw,
+		bod.KW,
+		bod.SW,
 		&bo,
-		bod.sel,
-		bod.sel.ID(),
+		bod.Sel,
+		bod.Sel.ID(),
 		path.LibrariesCategory)
 
 	bID := bo.Results.BackupID
@@ -390,32 +391,32 @@ func (suite *GroupsBackupIntgSuite) TestBackup_Run_groupsExtensions() {
 		ctx,
 		bID,
 		tenID,
-		bod.sel.ID(),
+		bod.Sel.ID(),
 		svc,
 		ws,
-		bod.kms,
-		bod.sss)
+		bod.KMS,
+		bod.SSS)
 	deeTD.CheckBackupDetails(
 		t,
 		ctx,
 		bID,
 		ws,
-		bod.kms,
-		bod.sss,
+		bod.KMS,
+		bod.SSS,
 		expectDeets,
 		false)
 
 	// Check that the extensions are in the backup
 	for _, ent := range deets.Entries {
 		if ent.Folder == nil {
-			verifyExtensionData(t, ent.ItemInfo, path.GroupsService)
+			VerifyExtensionData(t, ent.ItemInfo, path.GroupsService)
 		}
 	}
 }
 
 type GroupsBackupNightlyIntgSuite struct {
 	tester.Suite
-	its intgTesterSetup
+	its IntgTesterSetup
 }
 
 func TestGroupsBackupNightlyIntgSuite(t *testing.T) {
@@ -427,30 +428,30 @@ func TestGroupsBackupNightlyIntgSuite(t *testing.T) {
 }
 
 func (suite *GroupsBackupNightlyIntgSuite) SetupSuite() {
-	suite.its = newIntegrationTesterSetup(suite.T())
+	suite.its = NewIntegrationTesterSetup(suite.T())
 }
 
 func (suite *GroupsBackupNightlyIntgSuite) TestBackup_Run_groupsVersion9MergeBase() {
-	sel := selectors.NewGroupsBackup([]string{suite.its.group.ID})
+	sel := selectors.NewGroupsBackup([]string{suite.its.Group.ID})
 	sel.Include(
 		selTD.GroupsBackupLibraryFolderScope(sel),
 		selTD.GroupsBackupChannelScope(sel))
 
-	runMergeBaseGroupsUpdate(suite, sel.Selector, false)
+	RunMergeBaseGroupsUpdate(suite, sel.Selector, false)
 }
 
 func (suite *GroupsBackupNightlyIntgSuite) TestBackup_Run_groupsVersion9AssistBases() {
-	sel := selectors.NewGroupsBackup([]string{suite.its.group.ID})
+	sel := selectors.NewGroupsBackup([]string{suite.its.Group.ID})
 	sel.Include(
 		selTD.GroupsBackupLibraryFolderScope(sel),
 		selTD.GroupsBackupChannelScope(sel))
 
-	runDriveAssistBaseGroupsUpdate(suite, sel.Selector, false)
+	RunDriveAssistBaseGroupsUpdate(suite, sel.Selector, false)
 }
 
 type GroupsRestoreNightlyIntgSuite struct {
 	tester.Suite
-	its intgTesterSetup
+	its IntgTesterSetup
 }
 
 func TestGroupsRestoreIntgSuite(t *testing.T) {
@@ -462,20 +463,20 @@ func TestGroupsRestoreIntgSuite(t *testing.T) {
 }
 
 func (suite *GroupsRestoreNightlyIntgSuite) SetupSuite() {
-	suite.its = newIntegrationTesterSetup(suite.T())
+	suite.its = NewIntegrationTesterSetup(suite.T())
 }
 
 func (suite *GroupsRestoreNightlyIntgSuite) TestRestore_Run_groupsWithAdvancedOptions() {
-	sel := selectors.NewGroupsBackup([]string{suite.its.group.ID})
+	sel := selectors.NewGroupsBackup([]string{suite.its.Group.ID})
 	sel.Include(selTD.GroupsBackupLibraryFolderScope(sel))
 	sel.Filter(sel.Library("documents"))
-	sel.DiscreteOwner = suite.its.group.ID
+	sel.DiscreteOwner = suite.its.Group.ID
 
-	runDriveRestoreWithAdvancedOptions(
+	RunDriveRestoreWithAdvancedOptions(
 		suite.T(),
 		suite,
-		suite.its.ac,
+		suite.its.AC,
 		sel.Selector,
-		suite.its.group.RootSite.DriveID,
-		suite.its.group.RootSite.DriveRootFolderID)
+		suite.its.Group.RootSite.DriveID,
+		suite.its.Group.RootSite.DriveRootFolderID)
 }
